app/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
pgvector
google-generativeai
python-dotenv
requests
